refactor: move tr_log_level to types.h
